requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dataclasses-json==0.5.14
dnspython==2.4.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.279
langsmith==0.0.33
loguru==0.7.0
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
pinecone-client==2.2.2
pydantic==2.3.0
pydantic_core==2.6.3
pypdf==3.15.4
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wikipedia==1.4.0
yarl==1.9.2
